Factor out Arabic implementation
