+++ .github/workflows/python.yml
@@ -25,18 +25,18 @@ jobs:
         key: ${{ env.pythonLocation }}-${{ hashFiles('python/setup.py') }}-${{ hashFiles('python/requirements.txt') }}
 
     - name: Install requirements
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         pip install --upgrade --upgrade-strategy eager -r requirements.txt -e .
 
     - name: Download PyTorch model
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         curl -sSL https://github.com/secryst/rababa-models/releases/download/0.1/2000000-snapshot.pt \
           -o log_dir/CA_MSA.base.cbhg/models/2000000-snapshot.pt
 
     - name: Run diacriticization
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         python diacritize.py --model_kind "cbhg" --config config/cbhg.yml --text 'قطر'
 
@@ -59,12 +59,12 @@ jobs:
         key: ${{ env.pythonLocation }}-${{ hashFiles('python/setup.py') }}-${{ hashFiles('python/requirements.txt') }}
 
     - name: Install requirements
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         pip install --upgrade --upgrade-strategy eager -r requirements.txt -e .
 
     - name: Prepare dataset
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         mkdir -p data/CA_MSA
         touch data/CA_MSA/{eval,train,test}.csv
@@ -76,6 +76,6 @@ jobs:
         for d in `ls rababa-tashkeela-1.0/tashkeela_test/*`; do cat $d >> CA_MSA/test.csv; done
 
     - name: Try training (WIP)
-      working-directory: ./python
+      working-directory: ./python/arabic
       run: |
         python train.py --model "cbhg" --config config/test_cbhg.yml
