+++ data/example_arabic.txt
