+++ exe/rababa
@@ -8,9 +8,15 @@ require_relative "../lib/rababa"
 
 parser = Rababa.parser
 
-config_path = parser.has_key?(:config) ? parser[:config] : "config/model.yml"
+config_path = parser[:config]
 
-diacritizer = Rababa::Diacritizer.new(parser[:model_path], YAML.load_file(config_path))
+case parser[:language]
+when "arabic"
+  config_path ||= "config/model.yml"
+  diacritizer_class = Rababa::Arabic::Diacritizer
+end
+
+diacritizer = diacritizer_class.new(parser[:model_path], YAML.load_file(config_path))
 
 if parser.has_key?(:text)
   # run diacritization text if has :text
