+++ lib/rababa/arabic.rb
@@ -1,0 +1,4 @@
+require "rababa/arabic/constants"
+require "rababa/arabic/diacritizer"
+require "rababa/arabic/encoders"
+require "rababa/arabic/reconciler"
