+++ python/arabic/config_manager.py
