+++ python/arabic/log_dir/README.adoc
