+++ python/arabic/requirements.txt
