+++ python/arabic/setup.py
