+++ python/arabic/tester.py
