+++ python/arabic/train.py
