+++ python/arabic/trainer.py
