+++ python/arabic/util/constants.py
