+++ python/arabic/util/utils.py
